Split authorities-noting-inherent into 2 crates
