+++ container-chain-pallets/authorities-noting/src/mock.rs
@@ -272,7 +272,7 @@ where {
                 let inherent_data = {
                     let mut inherent_data = InherentData::default();
                     let system_inherent_data =
-                        ccp_authorities_noting_inherent::ContainerChainAuthoritiesInherentData {
+                        ccp_authorities_noting_inherent_core::ContainerChainAuthoritiesInherentData {
                             relay_chain_state,
                             orchestrator_chain_state: self
                                 .orchestrator_storage_proof
@@ -281,7 +281,7 @@ where {
                         };
                     inherent_data
                         .put_data(
-                            ccp_authorities_noting_inherent::INHERENT_IDENTIFIER,
+                            ccp_authorities_noting_inherent_core::INHERENT_IDENTIFIER,
                             &system_inherent_data,
                         )
                         .expect("failed to put VFP inherent");
