+++ container-chain-primitives/authorities-noting-inherent/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "ccp-authorities-noting-inherent"
 authors = { workspace = true }
-description = "authorities-noting-inherent primitives"
+description = "client-side authorities-noting-inherent primitives"
 edition = "2021"
 license = "GPL-3.0-only"
 version = "0.1.0"
@@ -15,6 +15,7 @@ parity-scale-codec = { workspace = true, features = [ "derive", "max-encoded-len
 scale-info = { workspace = true }
 tracing = { workspace = true, optional = true }
 
+ccp-authorities-noting-inherent-core = { workspace = true }
 dc-orchestrator-chain-interface = { workspace = true, optional = true }
 dp-collator-assignment = { workspace = true, optional = true }
 dp-core = { workspace = true, optional = true }
@@ -53,6 +54,7 @@ std = [
 	"cumulus-primitives-core/std",
 	"cumulus-primitives-parachain-inherent/std",
 	"cumulus-relay-chain-interface",
+	"ccp-authorities-noting-inherent-core/std",
 	"dc-orchestrator-chain-interface",
 	"dp-collator-assignment/std",
 	"dp-core/std",
