+++ container-chain-primitives/authorities-noting-inherent/src/tests.rs
@@ -16,7 +16,6 @@
 
 use {
     super::*,
-    crate::ContainerChainAuthoritiesInherentData,
     async_trait::async_trait,
     cumulus_primitives_core::{
         relay_chain::{
@@ -33,6 +32,7 @@ use {
     dp_core::{well_known_keys, Header as OrchestratorHeader},
     futures::Stream,
     nimbus_primitives::NimbusId,
+    parity_scale_codec::Encode,
     polkadot_overseer::Handle,
     polkadot_primitives::vstaging::CommittedCandidateReceiptV2 as CommittedCandidateReceipt,
     sc_client_api::{HeaderBackend, StorageKey, StorageProvider},
@@ -409,18 +409,17 @@ async fn test_orchestrator_inherent_insertion() {
         .unwrap();
 
     // get latest header info
-    let latest_header_info =
-        ContainerChainAuthoritiesInherentData::get_latest_orchestrator_head_info(
-            relay_header.hash(),
-            &relay_chain_interface,
-            orch_para_id.into(),
-        )
-        .await;
+    let latest_header_info = get_latest_orchestrator_head_info(
+        relay_header.hash(),
+        &relay_chain_interface,
+        orch_para_id.into(),
+    )
+    .await;
 
     // assert creation went well
     assert_eq!(latest_header_info, Some(orchestrator_header));
 
-    let created = ContainerChainAuthoritiesInherentData::create_at(
+    let created = create_at(
         relay_header.hash(),
         &relay_chain_interface,
         &orchestrator_chain_interface,
@@ -480,18 +479,17 @@ async fn test_header_not_present_error() {
         .unwrap();
 
     // get latest header info, but for another paraId
-    let latest_header_info =
-        ContainerChainAuthoritiesInherentData::get_latest_orchestrator_head_info(
-            relay_header.hash(),
-            &relay_chain_interface,
-            (orch_para_id + 1).into(),
-        )
-        .await;
+    let latest_header_info = get_latest_orchestrator_head_info(
+        relay_header.hash(),
+        &relay_chain_interface,
+        (orch_para_id + 1).into(),
+    )
+    .await;
 
     // assert creation went well
     assert_eq!(latest_header_info, None);
 
-    let created = ContainerChainAuthoritiesInherentData::create_at(
+    let created = create_at(
         relay_header.hash(),
         &relay_chain_interface,
         &orchestrator_chain_interface,
@@ -521,17 +519,16 @@ async fn test_head_data_not_decodable_error() {
         .unwrap();
 
     // get latest header info, but cannot since head data does not decode
-    let latest_header_info =
-        ContainerChainAuthoritiesInherentData::get_latest_orchestrator_head_info(
-            relay_header.hash(),
-            &relay_chain_interface,
-            (orch_para_id).into(),
-        )
-        .await;
+    let latest_header_info = get_latest_orchestrator_head_info(
+        relay_header.hash(),
+        &relay_chain_interface,
+        (orch_para_id).into(),
+    )
+    .await;
 
     assert_eq!(latest_header_info, None);
 
-    let created = ContainerChainAuthoritiesInherentData::create_at(
+    let created = create_at(
         relay_header.hash(),
         &relay_chain_interface,
         &orchestrator_chain_interface,
@@ -561,17 +558,16 @@ async fn test_header_not_decodable() {
         .unwrap();
 
     // get latest header info, but cannot since header does not decode
-    let latest_header_info =
-        ContainerChainAuthoritiesInherentData::get_latest_orchestrator_head_info(
-            relay_header.hash(),
-            &relay_chain_interface,
-            (orch_para_id).into(),
-        )
-        .await;
+    let latest_header_info = get_latest_orchestrator_head_info(
+        relay_header.hash(),
+        &relay_chain_interface,
+        (orch_para_id).into(),
+    )
+    .await;
 
     assert_eq!(latest_header_info, None);
 
-    let created = ContainerChainAuthoritiesInherentData::create_at(
+    let created = create_at(
         relay_header.hash(),
         &relay_chain_interface,
         &orchestrator_chain_interface,
